+++ app/config/collections.php
@@ -1560,6 +1560,15 @@ $collections = [
                 'required' => false,
                 'array' => false,
             ],
+            [
+                '$collection' => Database::SYSTEM_COLLECTION_RULES,
+                'label' => 'Async',
+                'key' => 'async',
+                'type' => Database::SYSTEM_VAR_TYPE_NUMERIC,
+                'default' => '',
+                'required' => false,
+                'array' => false,
+            ]
         ],
     ],
     Database::SYSTEM_COLLECTION_TAGS => [
@@ -1613,7 +1622,7 @@ $collections = [
                 'default' => '',
                 'required' => false,
                 'array' => false,
-            ],
+            ]
         ],
     ],
     Database::SYSTEM_COLLECTION_EXECUTIONS => [